pyproject.toml
[tool.poetry]
name = "nautobot-weather"
version = "0.1.0"
description = "Nautobot Weather"
authors = ["Andrew Bates <abates@omeganetserv.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/abates/nautobot-plugin-weather"
repository = "https://github.com/abates/nautobot-plugin-weather"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "LICENSE",
    "README.md",
]
packages = [
    { include = "nautobot_weather" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
# Used for local development
nautobot = { version = "^1.6.0", optional = true }
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
django-debug-toolbar = "*"
# we need to pin flake8 because of package dependencies that cause it to downgrade and
# therefore cause issues with linting since older versions do not take .flake8 as config
flake8 = "^3.9.2"
invoke = "*"
ipython = "*"
pydocstyle = "*"
pylint = "*"
pylint-django = "*"
pylint-nautobot = "*"
pytest = "*"
python-dotenv = "^0.21.1"
yamllint = "*"
Markdown = "*"
toml = "*"
# Rendering docs to HTML
mkdocs = "1.4.3"
# Material for MkDocs theme
mkdocs-material = "9.1.15"
# Render custom markdown for version added/changed/remove notes
mkdocs-version-annotations = "1.0.0"
# Automatic documentation from sources, for MkDocs
mkdocstrings = "0.22.0"         # Last version with python 3.7 support
mkdocstrings-python = "1.1.2"   # Last version with python 3.7 support
griffe = "0.30.1"               # Last version with python 3.7 support

[tool.poetry.extras]
nautobot = ["nautobot"]

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django, pylint_nautobot"
ignore=".venv"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.pylint-nautobot]
supported_nautobot_versions = [
    "1.6.0"
]

[tool.pydocstyle]
convention = "google"
inherit = false
match = "(?!__init__).*\\.py"
match-dir = "(?!tests|migrations|development)[^\\.].*"
# D212 is enabled by default in google convention, and complains if we have a docstring like:
# """
# My docstring is on the line after the opening quotes instead of on the same line as them.
# """
# We've discussed and concluded that we consider this to be a valid style choice.
add_ignore = "D212"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"
